small fixes

+++ src/logic/PlayerController.h
@@ -37,7 +37,7 @@ class PlayerController {
 	void updateInteraction();
 public:
 	static glm::vec3 selectedBlockPosition;
-	static glm::vec3 selectedBlockNormal;
+	static glm::ivec3 selectedBlockNormal;
 	static glm::vec3 selectedPointPosition;
 	static int selectedBlockId;
 	static int selectedBlockStates;
